Big refactor to merge the Classifier and CaseChat classes and get rid of the needlessly complex way messages are generated from prompts
