+++ langchainlaw/prompts.py
@@ -1,12 +1,7 @@
 from dataclasses import dataclass, field
-import yaml
 import json
 import random
 import re
-import pandas as pd
-
-from langchainlaw.generate_prompts import assemble_complete_yaml
-from langchain.schema import HumanMessage, SystemMessage
 
 JSON_QUOTE_RE = re.compile("```json(.*)```")
 
@@ -175,164 +170,3 @@ class CasePrompt:
         """Raise a PromptException if the config is invalid"""
         if self.return_type == "json_multiple" and not self.fields:
             raise PromptException("json_multiple needs a fields section")
-
-
-# Should this also be a dataclass?
-
-
-class CaseChat:
-    def __init__(self):
-        self._system = None
-        self._judgment = None
-        self._prompt_judgment = None
-        self._judgment_template = None
-        self._prompts = {}
-        self._prompt_names = []
-
-    @property
-    def system(self):
-        return self._system
-
-    @property
-    def prompt_names(self):
-        return self._prompt_names
-
-    @property
-    def judgment(self):
-        return self._judgment
-
-    @judgment.setter
-    def judgment(self, v):
-        self._judgment = v
-        self._prompt_judgment = self._judgment_template.format(judgment=json.dumps(v))
-
-    def prompt(self, name):
-        return self._prompts.get(name, None)
-
-    def assemble_yaml_from_spreadsheet(self, input_file):
-        output_file = input_file.split(".")[0] + ".yaml"
-        assemble_complete_yaml(input_file, output_file)
-        return output_file
-
-    def load_yaml(self, input_file):
-        output_file = self.assemble_yaml_from_spreadsheet(input_file)
-        with open(output_file, "r") as fh:
-            prompt_cf = yaml.load(fh, Loader=yaml.Loader)
-            self._system = prompt_cf["system"]
-            self._prompt_judgment_template = prompt_cf["intro"]
-            for p in prompt_cf["prompts"]:
-                self.add_prompt(
-                    p["name"],
-                    p["prompt"],
-                    p.get("return_type", "text"),
-                    p.get("fields", None),
-                    p.get("repeats", 1),
-                )
-            valid = True
-            for prompt in self.next_prompt():
-                try:
-                    prompt.validate()
-                except PromptException as e:
-                    print(f"Prompt '{prompt.name}' config error: {e}")
-                    valid = False
-        return valid
-
-    def load(self, spreadsheet):
-        """Load the prompts, system prompt and intro template from spreadsheet"""
-
-        system = pd.read_excel(spreadsheet, sheet_name="system")
-        # extract the first row from column System
-        self._system = system["System"][0]
-
-        intro = pd.read_excel(spreadsheet, sheet_name="intro")
-        self._judgment_template = intro["Intro"][0]
-        self.load_prompt_sheet(spreadsheet)
-
-    def load_prompt_sheet(self, spreadsheet):
-        prompts = pd.read_excel(spreadsheet, sheet_name="prompts", dtype=str).fillna("")
-
-        first_row = None
-        fields = []
-        for _, row in prompts.iterrows():
-            if row["return_type"]:
-                if first_row is not None:
-                    self.add_row(first_row, fields)
-                first_row = row[:]
-                fields = []
-            fields.append(
-                CasePromptField(
-                    field=row["fields"],
-                    question=row["question_description"],
-                    example_response=row["example"],
-                )
-            )
-
-        if first_row is not None:
-            self.add_row(first_row, fields)
-
-    def add_row(self, row, fields):
-        repeats = 1
-        if row["return_type"] == "json_multiple":
-            if row["repeats"]:
-                try:
-                    repeats = int(row["repeats"])
-                except ValueError:
-                    raise PromptException("'repeats' must be an integer")
-        self.add_prompt(
-            CasePrompt(
-                name=row["Prompt_name"],
-                question=row["prompt_question"],
-                return_instruction=row["return_instruction"],
-                return_type=row["return_type"],
-                additional_instruction=row["additional_instruction"],
-                fields=fields,
-                repeats=repeats,
-            )
-        )
-
-    def debug(self):
-        print("Debugging prompts")
-        for name in self.prompt_names:
-            prompt = self.prompt(name)
-            print(f"\nprompt {name}")
-            print(prompt.prompt)
-            print(prompt.return_type)
-            print(prompt.fields)
-
-    def add_prompt_old(self, name, prompt, return_type, fields, repeats):
-        if name in self._prompts:
-            raise ValueError(f"Prompt with name {name} already defined")
-        self._prompt_names.append(name)
-        self._prompts[name] = CasePrompt(
-            name=name,
-            prompt=prompt,
-            return_type=return_type,
-            fields=fields,
-            repeats=repeats,
-        )
-
-    def add_prompt(self, prompt):
-        """New version where you pass in a CasePrompt"""
-        if prompt.name in self._prompts:
-            raise ValueError(f"Prompt with name {prompt.name} already exists")
-        self._prompt_names.append(prompt.name)
-        self._prompts[prompt.name] = prompt
-
-    def start_chat(self):
-        return SystemMessage(content=self.system)
-
-    def next_prompt(self):
-        for prompt_name in self._prompt_names:
-            yield self._prompts[prompt_name]
-
-    def make_message(self, prompt):
-        """Builds the complete prompt from the JSON-encoded judgment and
-        the prompt questions (which also will include examples for the LLM to
-        return)"""
-        if self._prompt_judgment is not None:
-            content = self._prompt_judgment + prompt.prompt
-            return HumanMessage(content=content)
-        else:
-            raise PromptException(
-                "Need to set the judgment with judgment() before make_message()"
-            )
